finished implementing the Author Crud and its test suite

+++ Sami_Archive/Models/EFGenreRepository.cs
@@ -30,6 +30,7 @@ namespace Sami_Archive.Models
         public async Task DeleteGenreAsync(long GenreID)
         {
             var genre = await _context.Genres.FindAsync(GenreID);
+            if (genre == null) { return ; }
             _context.Genres.Remove(genre);
             await _context.SaveChangesAsync();
         }
